requirements.txt
streamlit
datasets
pandas
requests
python-dotenv
